test: isolate closed-window regressions from sibling fixes
